# api-restful/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API about the Marvel Cinematic Universe (MCU)",
        "title": "Marvel API",
        "contact": {
            "name": "Kyrex",
            "url": "github.com/kyrex23"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/heroes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Heroes"
                ],
                "summary": "Get the list of heroes",
                "responses": {
                    "200": {
                        "description": "List of heroes retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.heroResponse"
                            }
                        }
                    }
                }
            }
        },
        "/heroes/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Heroes"
                ],
                "summary": "Get a hero by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hero ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hero found successfully",
                        "schema": {
                            "$ref": "#/definitions/handler.heroResponse"
                        }
                    },
                    "404": {
                        "description": "Hero not found"
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.heroResponse": {
            "type": "object",
            "properties": {
                "actual_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}